# swagger.yaml
openapi: 3.0.0
info:
  title: 'ProviEmplea API'
  description: 'API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080/api'
    description: 'Servidor de desarrollo local'
paths:
  /admin/contactos:
    get:
      tags:
        - Administración
      summary: 'Listar contactos solicitados'
      operationId: getContactosSolicitados
      parameters:
        -
          name: estado
          in: query
          required: false
          schema:
            type: string
            enum:
              - pendiente
              - contactado
              - entrevista
              - seleccionado
              - no-seleccionado
              - proceso-cerrado
      responses:
        '200':
          description: 'Listado exitoso'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactoSolicitado'
    post:
      tags:
        - Administración
      summary: 'Registrar solicitud de contacto'
      description: 'Una empresa solicita contactar a un talento. No puede existir una solicitud activa previa.'
      operationId: crearContactoSolicitado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactoSolicitadoInput'
      responses:
        '201':
          description: 'Contacto registrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactoSolicitado'
        '409':
          description: 'Ya existe una solicitud activa'
        '422':
          description: 'Errores de validación'
  '/admin/contactos/{id}/estado':
    patch:
      tags:
        - Administración
      summary: 'Actualizar estado de contacto'
      description: 'Cambia el estado del proceso. Las fechas se registran automáticamente según el estado.'
      operationId: actualizarEstadoContacto
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - estado
              properties:
                estado:
                  type: string
                  enum: [pendiente, contactado, entrevista, seleccionado, no-seleccionado, proceso-cerrado]
                notas_admin:
                  type: string
                  nullable: true
              type: object
      responses:
        '200':
          description: 'Estado actualizado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactoSolicitado'
        '404':
          description: 'Contacto no encontrado'
...
  /empresas:
    get:
      tags:
        - Empresas
      summary: 'Listar empresas validadas'
      operationId: getEmpresas
      parameters:
        -
          name: tipo_empresa
          in: query
          required: false
          schema:
            type: string
            enum:
              - contratacion-directa
              - est
              - outsourcing
      responses:
        '200':
          description: 'Listado exitoso'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Empresa'
    post:
      tags:
        - Empresas
      summary: 'Registrar nueva empresa'
      operationId: createEmpresa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpresaInput'
      responses:
        '201':
          description: 'Empresa creada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empresa'
        '422':
          description: 'Errores de validación'
  '/empresas/{id}':
    get:
      tags:
        - Empresas
      summary: 'Obtener empresa por ID'
      operationId: getEmpresa
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Empresa encontrada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empresa'
        '404':
          description: 'No encontrada'
    put:
      tags:
        - Empresas
      summary: 'Actualizar empresa'
      operationId: updateEmpresa
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpresaInput'
      responses:
        '200':
          description: 'Empresa actualizada'
        '404':
          description: 'No encontrada'
    delete:
      tags:
        - Empresas
      summary: 'Desactivar empresa'
      description: 'Desactiva el perfil sin eliminarlo de la base de datos.'
      operationId: deleteEmpresa
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Empresa desactivada'
        '404':
          description: 'No encontrada'
  '/empresas/{id}/validar':
    patch:
      tags:
        - Empresas
      summary: 'Validar empresa (solo administración)'
      operationId: validarEmpresa
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Empresa validada'
        '404':
          description: 'No encontrada'
...
components:
  schemas:
    ContactoSolicitadoInput:
      title: 'Contacto Solicitado Input'
      required:
        - empresa_id
        - persona_id
      properties:
        empresa_id:
          type: integer
          example: 1
        persona_id:
          type: integer
          example: 1
        notas_admin:
          type: string
          nullable: true
      type: object
    ContactoSolicitado:
      title: 'Contacto Solicitado'
      description: 'Solicitud de contacto entre empresa y talento'
      properties:
        id:
          type: integer
          example: 1
        empresa_id:
          type: integer
          example: 1
        persona_id:
          type: integer
          example: 1
        estado:
          type: string
          example: pendiente
          enum:
            - pendiente
            - contactado
            - entrevista
            - seleccionado
            - no-seleccionado
            - proceso-cerrado
        notas_admin:
          type: string
          nullable: true
        fecha_contacto:
          type: string
          format: date-time
          nullable: true
        fecha_entrevista:
          type: string
          format: date-time
          nullable: true
        fecha_resultado:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    EmpresaInput:
      title: 'Empresa Input'
      description: 'Datos para crear o actualizar una empresa'
      required:
        - nombre_empresa
        - rut_empresa
        - email
        - tipo_empresa
        - contacto_nombre
        - contacto_email
      properties:
        nombre_empresa:
          type: string
          example: 'TechCorp SpA'
        rut_empresa:
          type: string
          example: 76123456-7
        email:
          type: string
          format: email
          example: rrhh@techcorp.cl
        logo_url:
          type: string
          format: url
          nullable: true
        rubro:
          type: string
          example: Tecnología
          nullable: true
        tipo_empresa:
          type: string
          enum:
            - contratacion-directa
            - est
            - outsourcing
        presentacion:
          type: string
          nullable: true
        beneficios:
          type: array
          items:
            type: string
          nullable: true
        contacto_nombre:
          type: string
          example: 'Ana López'
        contacto_email:
          type: string
          format: email
          example: ana@techcorp.cl
        contacto_telefono:
          type: string
          nullable: true
      type: object
    Empresa:
      title: Empresa
      description: 'Modelo completo de empresa empleadora'
      required:
        - nombre_empresa
        - rut_empresa
        - email
        - tipo_empresa
        - contacto_nombre
        - contacto_email
      properties:
        id:
          type: integer
          example: 1
        nombre_empresa:
          type: string
          example: 'TechCorp SpA'
        rut_empresa:
          type: string
          example: 76123456-7
        email:
          type: string
          format: email
          example: rrhh@techcorp.cl
        logo_url:
          type: string
          format: url
          nullable: true
        rubro:
          type: string
          example: Tecnología
          nullable: true
        tipo_empresa:
          type: string
          enum:
            - contratacion-directa
            - est
            - outsourcing
        presentacion:
          type: string
          nullable: true
        beneficios:
          type: array
          items:
            type: string
          example:
            - 'Seguro complementario'
            - 'Trabajo remoto'
          nullable: true
        contacto_nombre:
          type: string
          example: 'Ana López'
        contacto_email:
          type: string
          format: email
          example: ana@techcorp.cl
        contacto_telefono:
          type: string
          example: '+56912345678'
          nullable: true
        validado:
          type: boolean
          example: false
        activo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
...